// src/app/pages/star-rail/models/star-rail-enums.ts
export enum RelicTypeEnum {
  HEAD = 'Head',
  HANDS = 'Hands',
  BODY = 'Body',
  FEET = 'Feet',
  PLANAR_SPHERE = 'Planar Sphere',
  LINK_ROPE = 'Link Rope'
}

export enum MainStatsEnum {
  SPD = 'SPD',
  FLAT_HP = 'HP',
  FLAT_ATK = 'ATK',
  PERC_HP = 'HP%',
  PERC_ATK = 'ATK%',
  PERC_DEF = 'DEF%',
  BREAK_EFFECT = 'Break Effect',
  EFFECT_HIT_RATE = 'Effect Hit Rate',
  ENERGY_REG = 'Energy Regeneration Rate',
  OUTGOING_HEAL = 'Outgoing Healing Boost',
  PHYS_DMG = 'Physical Damage Boost',
  FIRE_DMG = 'Fire Damage Boost',
  ICE_DMG = 'Ice Damage Boost',
  WIND_DMG = 'Wind Damage Boost',
  LIGHT_DMG = 'Lightning Damage Boost',
  QUANT_DMG = 'Quantum Damage Boost',
  IMG_DMG = 'Imaginary Damage Boost',
  CRIT_RATE = 'CRIT Rate',
  CRIT_DMG = 'CRIT DMG'
}